Some diff

+++ dmarc-convertor.sh
@@ -12,14 +12,15 @@
 # 3. Convert dmarc xml files to line oriented format for splunk
 #
 
-ROOT='/'
-DMARC_ROOT="${ROOT}/var/dmarc-report-processor"
+ROOT='/usr/local/dmarc-report-processor'
+DMARC_ROOT="${ROOT}/var"
 ATTACH="${DMARC_ROOT}/attach_raw"
 XML="${DMARC_ROOT}/dmarc_xml"
 DMARC_SPLUNK="${DMARC_ROOT}/dmarc_splunk"
 
 os=`uname`
-ydate=`date -d "yesterday 13:00 " '+%d-%h-%Y'`
+ydate=`date -d "yesterday 00:00 " '+%d-%h-%Y'`
+#ydate=`date -d "1 week ago 13:00 " '+%d-%h-%Y'`
 if [ "$os" == "Darwin" ]
 then
   ydate=`date -v-1d +%d-%h-%Y`
@@ -55,7 +56,7 @@ EOF
 }
 
 OPTIND=1
-while getopts "hs:p:u:P:" opt; do
+while getopts "hs:p:u:P:c:" opt; do
 case "$opt" in
   h)
     dmarc_help
@@ -107,7 +108,7 @@ d_search="SINCE \"${ydate}\" BEFORE \"${tdate}\""
 #1
 echo "Step 1: Fetch dmarc reports from mailbox"
 echo "----------------------------------------"
-${ROOT}/bin/imap-client.py --attachmentsonly -s "${d_host}" -c "${d_cert}" --port "${d_port}"  -u "${d_user}" -o ${ATTACH}/${ydate} -f inbox --pwdfile "${d_pwd}" -S "${d_search}"
+${ROOT}/bin/imap-client.py --attachmentsonly -s "${d_host}" -c "${d_cert}" --port "${d_port}"  -u "${d_user}" -o ${ATTACH}/${ydate} -f ${RUAFOLDER} --pwdfile "${d_pwd}" -S "${d_search}"
 if [ "$?" -ne "0" ]
 then
   echo "Error: imap-client mail attachment fetch failed; exiting ..."
@@ -115,11 +116,19 @@ then
 fi
 
 #2
+shopt -s nullglob
+files=( "${ATTACH}/${ydate}"/*.zip )
+if [ "${#files[@]}" -eq "0" ]
+then
+        echo "No new reports found. Exiting ..."
+        exit 0
+fi
+
 echo "Step 2: Unzipping files"
 echo "-----------------------"
 mkdir "${XML}/${ydate}"
 rm -rf "${XML}/${ydate}/*" 2> /dev/null
-for f in "${ATTACH}/${ydate}"/*; do
+for f in "${files[@]}"; do
   echo "$f"
   extn="${f##*.}"
   if [ "$extn" == "zip" ]
